vertex-pdf-summary-example/requirements.txt
pypdf
vertexai
python-dotenv
google-generativeai